Use port value for transmission client

+++ modules/transmission/settings.go
@@ -12,7 +12,7 @@ type Settings struct {
 	host     string `help:"The address of the machine the Transmission daemon is running on"`
 	https    bool   `help:"Whether or not to connect to the host via HTTPS"`
 	password string `help:"The password for the Transmission user"`
-	port     int    `help:"The port to connect to the Transmission daemon on"`
+	port     uint16 `help:"The port to connect to the Transmission daemon on"`
 	url      string `help:"The RPC URI that the daemon is accessible at"`
 	username string `help:"The username of the Transmission user"`
 }
@@ -29,7 +29,7 @@ func NewSettingsFromYAML(name string, ymlConfig *config.Config, globalConfig *co
 		host:     ymlConfig.UString("host"),
 		https:    ymlConfig.UBool("https", false),
 		password: ymlConfig.UString("password"),
-		port:     ymlConfig.UInt("port", 9091),
+		port:     uint16(ymlConfig.UInt("port", 9091)),
 		url:      ymlConfig.UString("url", "/transmission/"),
 		username: ymlConfig.UString("username", ""),
 	}
